fix: prevent arithmetic overflow and truncation in cost tracking, replication, and tree operations (#540)
Replace wrapping arithmetic with saturating ops and checked conversions:
- L1: saturating_sub in epoch storage flag removal to prevent underflow
- L3: saturating_add/sub in OperationCost Add/AddAssign and paid_value_len
- L7: checked_add for bounds check and u32-to-usize truncation guard in replication
- L8: compute balance_factor in i16 and clamp to i8 to prevent wrapping
- L13: change len_of_values_at_path return type from u16 to usize
- L18: use u8::try_from with saturating fallback for key length
- L19: use u16::try_from with saturating fallback for ops count

+++ grovedb/src/query_result_type.rs
@@ -119,7 +119,7 @@ impl BTreeMapLevelResult {
 
 impl BTreeMapLevelResult {
     /// Returns the number of values at the given path within this result tree.
-    pub fn len_of_values_at_path(&self, path: &[&[u8]]) -> u16 {
+    pub fn len_of_values_at_path(&self, path: &[&[u8]]) -> usize {
         let mut current = self;
 
         // Traverse the path
@@ -139,7 +139,7 @@ impl BTreeMapLevelResult {
             }
         }
 
-        current.key_values.len() as u16
+        current.key_values.len()
     }
 }
